+++ ffi/dl_android.go
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: Apache-2.0
 // SPDX-FileCopyrightText: 2026 The Goffi Authors
 
-//go:build android && arm64
+//go:build android && arm64 && !goffi_static
 
 package ffi