+++ internal/fakecgo/android_abi_test.go
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: Apache-2.0
 // SPDX-FileCopyrightText: 2026 The Goffi Authors
 
-//go:build android && arm64 && !cgo
+//go:build android && arm64 && !cgo && !goffi_static
 
 package fakecgo