+++ internal/fakecgo/libcgo.go
@@ -2,7 +2,7 @@
 // SPDX-FileCopyrightText: 2022 The Ebitengine Authors
 // SPDX-FileCopyrightText: 2025-2026 Andrey Kolkov and GoGPU Contributors
 
-//go:build !cgo && (darwin || freebsd || linux || netbsd) && !android
+//go:build !cgo && (darwin || freebsd || linux || netbsd) && !android && !goffi_static
 
 package fakecgo