+++ internal/fakecgo/libcgo_freebsd.go
@@ -2,7 +2,7 @@
 // SPDX-FileCopyrightText: 2022 The Ebitengine Authors
 // SPDX-FileCopyrightText: 2025-2026 Andrey Kolkov and GoGPU Contributors
 
-//go:build !cgo
+//go:build !cgo && !goffi_static
 
 package fakecgo